rename files and improve performance
